requirements.txt
together==1.4.6
flask==3.0.0
python-dotenv==1.0.0
speechrecognition==3.10.0
mediapipe==0.10.21
matplotlib==3.8.0
six==1.16.0
opencv-python==4.9.0.80
numpy==1.26.0
gunicorn==21.2.0
google-cloud-texttospeech==2.16.0